refactor: rename "constants" folder to "mocks"
