+++ src/core/packages/chrome/navigation/src/__stories__/navigation.stories.tsx
@@ -16,7 +16,7 @@ import { Box } from '@kbn/core-chrome-layout-components/__stories__/box';
 import { APP_MAIN_SCROLL_CONTAINER_ID } from '@kbn/core-chrome-layout-constants';
 
 import { Navigation } from '../components/navigation';
-import { LOGO, PRIMARY_MENU_ITEMS, PRIMARY_MENU_FOOTER_ITEMS } from '../constants/observability';
+import { LOGO, PRIMARY_MENU_ITEMS, PRIMARY_MENU_FOOTER_ITEMS } from '../mocks/observability';
 import { NavigationStructure } from '../../types';
 
 const styles = ({ euiTheme }: UseEuiTheme) => css`
