+++ src/core/packages/chrome/navigation/src/mocks/observability.ts
